Move SavedColumnSettings
