+++ src/data/data-table/entity-service/RenameTableModal.js
@@ -3,7 +3,6 @@ import { Fragment, useState } from 'react';
 import { div, h } from 'react-hyperscript-helpers';
 import { ButtonPrimary, IdContainer, RadioButton, spinnerOverlay } from 'src/components/common';
 import { warningBoxStyle } from 'src/components/data/data-utils';
-import { allSavedColumnSettingsEntityTypeKey } from 'src/components/data/SavedColumnSettings';
 import { icon } from 'src/components/icons';
 import { ValidatedInput } from 'src/components/input';
 import Modal from 'src/components/Modal';
@@ -13,6 +12,8 @@ import { withErrorReporting } from 'src/libs/error';
 import Events from 'src/libs/events';
 import { FormLabel } from 'src/libs/forms';
 import * as Utils from 'src/libs/utils';
+
+import { allSavedColumnSettingsEntityTypeKey } from './SavedColumnSettings';
 
 export const RenameTableModal = ({
   onDismiss,
